# Docs - contract/openapi.yaml
openapi: 3.0.1
info:
  title: Homely API
  description: HTTP JSON dùng {statusCode,message,data}; DTO trong data dùng snake_case, VND là chuỗi số nguyên. Thanh toán và hoàn tiền là sandbox.
  version: 1.6.0
servers:
- description: Local backend; paths include /api/v1
  url: http://localhost:8080
paths:
...
  /api/v1/auth/verify-email:
    post:
      tags:
      - auth-controller
      summary: Verify email address
      operationId: verifyEmail
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OneTimeTokenRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    type: object
                    additionalProperties:
                      type: string
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
  /api/v1/auth/verify-email/resend:
    post:
      tags:
      - auth-controller
      summary: Resend email verification
      operationId: resendVerification
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    type: object
                    additionalProperties:
                      type: string
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
  /api/v1/auth/register:
    post:
      tags:
      - auth-controller
      summary: Register a new user
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/UserResponse'
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
  /api/v1/auth/refresh:
    post:
      tags:
      - auth-controller
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/LoginResponse'
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
  /api/v1/auth/logout:
    post:
      tags:
      - auth-controller
      summary: Logout user
      operationId: logout
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    type: object
                    nullable: true
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
  /api/v1/auth/login:
    post:
      tags:
      - auth-controller
      summary: Login by credential
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/LoginResponse'
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
  /api/v1/auth/enable-host:
    post:
      tags:
      - auth-controller
      summary: Enable host role for current user
      operationId: enableHost
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/UserResponse'
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
  /api/v1/auth/delete-account:
    get:
      tags:
      - auth-controller
      summary: Get current account deletion request
      operationId: deletionStatus
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/AccountDeletionResponse'
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
    post:
      tags:
      - auth-controller
      summary: Request account deletion
      operationId: requestDeletion
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeletionRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    type: object
                    additionalProperties:
                      type: string
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
...
  /api/v1/auth/account:
    get:
      tags:
      - auth-controller
      summary: Get user information
      operationId: getAccount
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/UserLogin'
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
...
components:
  schemas:
...
    OneTimeTokenRequest:
      required:
      - token
      type: object
      properties:
        token:
          maxLength: 512
          minLength: 43
          type: string
    RegisterRequest:
      required:
      - email
      - fullName
      - password
      type: object
      properties:
        email:
          type: string
        password:
          maxLength: 128
          minLength: 12
          type: string
        fullName:
          maxLength: 100
          minLength: 2
          type: string
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        fullName:
          type: string
        phone:
          type: string
        avatarUrl:
          type: string
        host:
          type: boolean
        emailVerified:
          type: boolean
        status:
          type: string
        createdAt:
          type: string
          format: date-time
    RefreshRequest:
      required:
      - refreshToken
      type: object
      properties:
        refreshToken:
          type: string
        installationId:
          type: string
        deviceName:
          type: string
    LoginResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserLogin'
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          format: int64
    UserLogin:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        fullName:
          type: string
        avatarUrl:
          type: string
        host:
          type: boolean
        emailVerified:
          type: boolean
        status:
          type: string
        roles:
          type: array
          items:
            type: string
    LoginRequest:
      required:
      - deviceName
      - email
      - installationId
      - password
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        installationId:
          type: string
        deviceName:
          maxLength: 100
          minLength: 1
          type: string
    DeletionRequest:
      required:
      - currentPassword
      type: object
      properties:
        currentPassword:
          type: string
        reason:
          maxLength: 1000
          minLength: 0
          type: string
...
    AccountDeletionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        status:
          type: string
        failureCode:
          type: string
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
...
    FieldErrorDTO:
      type: object
      properties:
        field:
          type: string
        reason:
          type: string
    ProblemDTO:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
        timestamp:
          type: string
          format: date-time
        field_errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDTO'
        trace_id:
          type: string
    RestResponseProblemDTO:
      required:
      - statusCode
      - message
      - data
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
          example: 500
        message:
          type: string
        data:
          $ref: '#/components/schemas/ProblemDTO'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...